Rework layouts and components for home page
