+++ app/View/Components/navigation/sidebar.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\View\Components\navigation;
+
+use Illuminate\View\Component;
+
+class sidebar extends Component
+{
+
+    public $tags;
+    /**
+     * Create a new component instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        //
+    }
+
+    /**
+     * Get the view / contents that represent the component.
+     *
+     * @return \Illuminate\Contracts\View\View|\Closure|string
+     */
+    public function render()
+    {
+        return view('components.navigation.sidebar');
+    }
+}
