+++ public/js/home/misc.js
