+++ resources/views/common/scripts.blade.php
@@ -1,0 +1,4 @@
+<script src="{{asset('js/jquery-3.6.1.min.js')}}"></script>
+<script src="{{asset('js/popper.min.js')}}"></script>
+<script src="{{asset('js/bootstrap.bundle.min.js')}}"></script>
+<script src="{{asset('js/masonry.pkgd.min.js')}}"></script>
