+++ resources/views/components/card/note.blade.php
