+++ resources/views/components/form/color-picker.blade.php
