--- resources/views/components/tag.blade.php
@@ -1,2 +1,0 @@
-<button class="btn card-tag bg-light rounded-pill text-center pt-0 mx-1 my-1 "><p class="mb-5">{{ $slot }}</p></button>
-
