+++ resources/views/components/tag/pill.blade.php
@@ -1,0 +1,1 @@
+<button class="btn card-tag bg-light rounded-pill text-center pt-0 mx-1 my-1 "><p class="mb-5">{{ $slot }}</p></button>
