+++ resources/views/layouts/simple.blade.php
@@ -1,0 +1,5 @@
+@extends('layouts.base')
+
+@section('content')
+
+@stop
